v2.0!

+++ scripts/build_site.py
@@ -6,6 +6,7 @@ import glob
 import re
 
 import list_to_list
+import print_cockatrice_file
 import print_draft_file
 import print_html_for_index
 import print_html_for_search
@@ -96,7 +97,7 @@ def removeStaleFiles(set_dir):
 		s_dir = os.path.join(set_dir, entry.name)
 		for set_entry in os.scandir(s_dir):
 			filename, file_extension = os.path.splitext(set_entry.name)
-			if set_entry.name not in filesToKeep and file_extension != '.json':
+			if set_entry.name not in filesToKeep and file_extension != '.json' and file_extension != '.xml':
 				if set_entry.is_dir():
 					shutil.rmtree(set_entry)
 				else:
@@ -106,6 +107,14 @@ def removeStaleFiles(set_dir):
 for entry in os.scandir('.'):
 	if '-spoiler' in entry.name:
 		os.remove(entry)
+
+#CE: auto-generate site-config.json
+repo_name = os.path.basename(os.getcwd())
+default_config = {
+	"base_url": f"https://{repo_name}"
+}
+with open(os.path.join('resources', 'site-config.json'), 'w', encoding='utf-8-sig') as f:
+	json.dump(default_config, f, indent=4)
 
 #F: first, get all the set codes
 set_codes = []
@@ -127,6 +136,10 @@ removeStaleFiles('sets')
 #CE: copy the entire custom tree
 portCustomFiles('custom', '')
 
+#CE: reload config in case it was overwritten by custom files
+import utils
+utils.load_config()
+
 #F: sort them
 set_codes.sort()
 
@@ -141,12 +154,20 @@ for code in set_codes:
 	set_dir = code + '-files'
 	with open(os.path.join('sets', code + '-files', code + '.json'), encoding='utf-8-sig') as f:
 		raw = json.load(f)
-	if 'draft_structure' in raw and not raw['draft_structure'] == 'none' and not os.path.isfile(os.path.join('custom', 'sets', code + '-files', code + '-draft.txt')):
+	if 'draft_structure' not in raw or not raw['draft_structure'] == 'none' and not os.path.isfile(os.path.join('custom', 'sets', code + '-files', code + '-draft.txt')):
 		try:
 			print_draft_file.generateFile(code)
 			print('Generated draft file for {0}.'.format(code))
 		except Exception as e:
 			print('Unable to generate draft file for {0}: {1}'.format(code, e))
+
+	# CE: Trice
+	if not os.path.isfile(os.path.join('custom', 'sets', code + '-files', code + '.xml')):
+		try:
+			print_cockatrice_file.generateFile(code)
+			print('Generated Cockatrice file for {0}.'.format(code))
+		except Exception as e:
+			print('Unable to generate Cockatrice file for {0}: {1}'.format(code, e))
 
 	#CE: this code is all for version history
 	if 'version' not in raw:
